requirements.txt
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
dotenv